+++ src/main/java/com/github/camilormz/finalreality/model/character/CharacterDomain.java
@@ -1,0 +1,15 @@
+package com.github.camilormz.finalreality.model.character;
+
+/**
+ * Enumeration of the domains that any character must have
+ * Understanding "domain" as a set of characteristics and behaviours of the character relative to
+ * the game mechanics
+ *
+ * In the current implementation, it defines if a character is playable or not (enemy).
+ *
+ * @author Camilo Ramírez Canales
+ */
+
+public enum CharacterDomain {
+    PLAYABLE, ENEMY
+}