+++ src/main/java/com/github/camilormz/finalreality/model/character/ICharacter.java
@@ -1,0 +1,34 @@
+package com.github.camilormz.finalreality.model.character;
+
+/**
+ * This represents a character from the game.
+ * A character can be controlled by the player or by the CPU (an enemy).
+ *
+ * @author Ignacio Slater Muñoz.
+ * @author Camilo Ramírez Canales.
+ */
+public interface ICharacter {
+
+  /**
+   * Sets a scheduled executor to make this character (thread) wait for {@code speed / 10}
+   * seconds before adding the character to the queue.
+   */
+  void waitTurn();
+
+  /**
+   * Gets the associated turn weight of the character in order to compute its enqueuing speed in
+   * the turns queue
+   */
+  int getTurnWeight();
+
+  /**
+   * Returns this character's name.
+   */
+  String getName();
+
+  /**
+   * Returns this character's domain
+   * @see CharacterDomain
+   */
+  CharacterDomain getCharacterDomain();
+}