+++ src/main/java/com/github/camilormz/finalreality/model/character/player/CharacterClass.java
@@ -1,0 +1,11 @@
+package com.github.camilormz.finalreality.model.character.player;
+
+/**
+ * Enumeration of the classes a player character may have.
+ *
+ * @author Ignacio Slater Muñoz.
+ * @author Camilo Ramírez Canales
+ */
+public enum CharacterClass {
+  KNIGHT, ENGINEER, THIEF, BLACK_MAGE, WHITE_MAGE
+}