+++ src/main/java/com/github/camilormz/finalreality/model/character/player/IMagicalCharacter.java
@@ -1,0 +1,10 @@
+package com.github.camilormz.finalreality.model.character.player;
+
+/**
+ * This represents a magical playable character from the game.
+ *
+ * @author Camilo Ramírez Canales.
+ */
+public interface IMagicalCharacter extends IPlayerCharacter {
+    // TODO: Here will be added the magical extra methods
+}