+++ src/main/java/com/github/camilormz/finalreality/model/weapon/WeaponType.java
@@ -1,0 +1,11 @@
+package com.github.camilormz.finalreality.model.weapon;
+
+/**
+ * Enumeration of all the weapon types.
+ *
+ * @author Ignacio Slater Muñoz.
+ * @author <Your name>
+ */
+public enum WeaponType {
+  SWORD, AXE, KNIFE, STAFF, BOW
+}