+++ src/test/java/com/github/camilormz/finalreality/model/character/EnemyTest.java
@@ -1,0 +1,9 @@
+package com.github.camilormz.finalreality.model.character;
+
+// TODO: This test IS totally broken and must be re-done ASAP
+
+class EnemyTest extends AbstractCharacterTest {
+
+  private static final String ENEMY_NAME = "Goblin";
+
+}