+++ src/test/java/com/github/camilormz/finalreality/model/weapon/AbstractWeaponTest.java
@@ -1,0 +1,21 @@
+package com.github.camilormz.finalreality.model.weapon;
+
+// TODO: This test is totally broken and must be re-done ASAP
+
+class AbstractWeaponTest {
+
+  private static final String AXE_NAME = "Test Axe";
+  private static final String STAFF_NAME = "Test Staff";
+  private static final String SWORD_NAME = "Test Sword";
+  private static final String BOW_NAME = "Test Bow";
+  private static final String KNIFE_NAME = "Test Knife";
+  private static final int DAMAGE = 15;
+  private static final int SPEED = 10;
+
+  private AbstractWeapon testAxe;
+  private AbstractWeapon testStaff;
+  private AbstractWeapon testSword;
+  private AbstractWeapon testBow;
+  private AbstractWeapon testKnife;
+
+}